+++ bc_obps/compliance/api/__init__.py
@@ -2,7 +2,7 @@
 from .router import router
 
 # Import new API structure
-from .compliance_report_version import get_compliance_report_versions_list
+from .compliance_report_versions import get_compliance_report_versions_list
 from ._compliance_report_version.compliance_report_version_id import get_compliance_report_version
 
 # from ._summaries._summary_id.issuance import get_compliance_summary_issuance
